requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.5
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6
ultralytics>=8.0.200
opencv-python-headless>=4.8.1.78
shapely>=2.0.2
pandas>=2.1.1
openai>=1.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Streamlit apps (admin_app, app_ui) — must be in root requirements for Streamlit Cloud
streamlit>=1.28.0
altair>=5.0.0
yt-dlp
psutil
